PR-3.12 PCS Copy migration + title gap + photo kebab removal
Three coordinated cleanups missed in PR-3.10:

1. Copy icon migrated from KickerRow top bar to CaptionBlock bottom
   action row as the leftmost action (next to WORDS counter, before
   Edit/AI/Proofread). Tap copies caption, long-press enters edit
   mode (canEdit only). Top bar collapses to X / chevrons / status
   / sidebar.

2. Title-to-caption gap bumped from pb-1 (4px) to pb-5 (20px) on the
   title wrapper in PCS.jsx — visually equals one caption line-height
   so the title gets room to breathe.

3. PhotoKebabMenu removed from the post-detail carousel inline mount.
   Lightbox kebab (topRightSlot) preserved with all five actions.
   Carousel now renders zero icons over the photo — only middle-bottom
   dots from PR-3.10.

Bumped: ?v= 20260425o → 20260425p (28 strings).
React bundle 0.9.23 → 0.9.24.
