+++ srtd-next/src/flows/pcs/components/KickerRow.jsx
@@ -1,14 +1,11 @@
 import React, { useEffect, useState } from 'react';
-import { X, PanelRight, ChevronLeft, ChevronRight, Copy } from 'lucide-react';
+import { X, PanelRight, ChevronLeft, ChevronRight } from 'lucide-react';
 import { STAGE_LABELS, STAGE_TOKEN } from '../utils/stage.js';
 import { pcsFlow } from '../index.js';
 import { useIsClient } from '../../../core/stores/appState.js';
 import { daysSince } from '../../plan/shared/dateUtils.js';
 import { usePcsFlowState } from '../flowStore.js';
 import { usePcsStore } from '../pcsStore.js';
-import { useLongPress } from '../../plan/hooks/useLongPress.js';
-import { toast } from '../../../core/bridges/toast.js';
-import { logClick } from '../../../core/bridges/logging.js';
 
 // Aging threshold: AgeBadge (plan/shared/AgeBadge.jsx) treats days >= 7 as
 // the red bucket. For the client-facing overdue capsule we render only when
@@ -26,47 +23,6 @@ export function KickerRow({ post, isAdmin, canMove, canEdit, onOpenSheet, onOpen
   const hasPrev = idx > 0;
   const hasNext = idx >= 0 && idx < contextList.length - 1;
   const navVisible = contextList && contextList.length > 1 && idx >= 0;
-
-  const copyLongPress = useLongPress({
-    enabled: !!canEdit,
-    threshold: 500,
-    onLongPress: () => {
-      try { usePcsStore.getState().requestCaptionEdit(); } catch (e) { /* noop */ }
-    },
-  });
-
-  async function copyCaption() {
-    const text = (post && post.caption) || '';
-    let ok = false;
-    try {
-      if (navigator.clipboard && navigator.clipboard.writeText) {
-        await navigator.clipboard.writeText(text);
-        ok = true;
-      }
-    } catch (e) { /* fall through to legacy path */ }
-    if (!ok) {
-      try {
-        const ta = document.createElement('textarea');
-        ta.value = text;
-        ta.style.position = 'fixed';
-        ta.style.top = '-1000px';
-        document.body.appendChild(ta);
-        ta.select();
-        document.execCommand('copy');
-        document.body.removeChild(ta);
-      } catch (e) { /* noop */ }
-    }
-    try { logClick('pcs_react_caption_copy', { len: text.length }); } catch (e) {}
-    toast('Copied', 'success');
-  }
-
-  function handleCopyClick(e) {
-    if (copyLongPress.onClick) {
-      copyLongPress.onClick(e);
-      if (e.defaultPrevented) return;
-    }
-    copyCaption();
-  }
 
   useEffect(() => {
     const onScroll = () => setCompressed(window.scrollY > 60);
@@ -185,17 +141,6 @@ export function KickerRow({ post, isAdmin, canMove, canEdit, onOpenSheet, onOpen
         </span>
       </div>
       <div className="flex items-center gap-0.5 flex-shrink-0 pr-2">
-        <button
-          type="button"
-          {...copyLongPress}
-          onClick={handleCopyClick}
-          onContextMenu={(e) => e.preventDefault()}
-          className="w-9 h-9 inline-flex items-center justify-center rounded-sm2 text-text-mid hover:text-text-loud active:bg-bg-2 active:scale-[0.96]"
-          style={{ transition: 'background 0.08s ease, color 0.1s ease, transform 0.08s ease' }}
-          aria-label="Copy caption"
-        >
-          <Copy size={16} strokeWidth={1.75} />
-        </button>
         <button
           onClick={onOpenSheet}
           className="w-9 h-9 inline-flex items-center justify-center rounded-sm2 text-text-mid hover:text-text-loud active:bg-bg-2 active:scale-[0.96]"
